fix: dev: Taiga #812 NTP uptime not updated correctly
ROOT CAUSE:
Because of high cpu utilization and lot of updates to OVSDB,
there exists a small window during which the background
daemon cannot get proper statistics/status from ntpd daemon.
Because of which the update to OVSDB also fails.
Here as part of the test, we were trying to configure
around 1024 loopback interface in one shot.

SOLUTION:
Background daemon script was handling error conditions where
it cannot get status info properly. So now we try and catch
such situations and flag an error to user that the daemon
is unable to synchronize information from NTPD -> OVSDB.

Change-Id: I785592b80f68e67354de624f096850b892b34a03
Signed-off-by: Srinivasan Srivatsan <srinivasan.srivatsan@hpe.com>